Extract ParallelChainImpl into its own file
