+++ src/common/parallel/parallel-task-scheduling.ts
@@ -1,0 +1,4 @@
+export interface IParallelTaskScheduling {
+    numberOfWorkers: number;
+    valuesPerWorker: number;
+}
